+++ js/util.js
@@ -1,0 +1,36 @@
+const getRandomPositiveInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomPositiveFloat = (a, b, digits = 5) => {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return +result.toFixed(digits);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const shuffle = (array) => {
+  const result = [...array];
+  for (let i = result.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [result[i], result[j]] = [result[j], result[i]];
+  }
+  return result;
+};
+
+const getRandomArray = (array) => shuffle(array).slice(0, getRandomPositiveInteger(0, array.length));
+
+const addLeadingZeros = (value, size) => {
+  let result = `${value}`;
+  while (result.length < size) {
+    result = `0${result}`;
+  }
+  return result;
+};
+
+export {getRandomPositiveInteger, getRandomPositiveFloat, getRandomArrayElement, shuffle, getRandomArray, addLeadingZeros};
